requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
numpy==1.19.4
pandas==1.1.5
psycopg2==2.7
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
